youtube_data_analysis/requirements.txt
mysql-connector-python
pandas
streamlit
google-api-python-client
